refactor: move header out of BasicPage
I'm currently trying to destroy the BasicPage class because it
has too many responsibilities and is overly rigid. In particular,
I want to stop modelling the stop page as just a basic page
and split it into two based on the two HTTP requests.

This commit just moves the header functionality out of BasicPage.
In a follow up commit I may just make BasicPage a wrapper around
the loading-and-reveal functionality.